# Conflicts:
#	python/sglang/srt/configs/model_config.py
#	python/sglang/srt/configs/utils.py
#	python/sglang/srt/layers/logits_processor.py
#	python/sglang/srt/managers/io_struct.py
#	python/sglang/srt/mem_cache/allocator.py
#	python/sglang/srt/mem_cache/memory_pool.py
#	python/sglang/srt/mem_cache/radix_cache.py
#	python/sglang/srt/model_executor/model_runner.py
#	python/sglang/srt/models/glm4_moe.py
#	python/sglang/srt/models/qwen3_next.py
#	python/sglang/srt/utils/hf_transformers_utils.py

+++ python/sglang/srt/speculative/eagle_worker.py
@@ -531,7 +531,7 @@ class EAGLEWorker(TpModelWorker):
         batch.spec_info = EagleDraftInput.create_idle_input(
             device=self.device,
             hidden_size=self.model_config.hidden_size,
-            dtype=self.model_config.dtype,
+            dtype=torch.bfloat16,
             topk=self.topk,
             capture_hidden_mode=CaptureHiddenMode.LAST,
         )
@@ -940,7 +940,7 @@ class EAGLEWorker(TpModelWorker):
             batch.spec_info = EagleDraftInput.create_idle_input(
                 device=self.device,
                 hidden_size=hidden_size,
-                dtype=self.model_config.dtype,
+                dtype=torch.bfloat16,
                 topk=self.topk,
                 capture_hidden_mode=CaptureHiddenMode.LAST,
             )
